requirements.txt
aiogram>=3.0.0
asyncpg>=0.27.0
redis[hiredis]>=4.5.0
aioredis>=2.0.0
APScheduler>=3.10.0
python-dotenv>=1.0.0
structlog>=23.0.0
babel>=2.12.0
setuptools>=65.0.0
pytest-asyncio>=0.21.0
ruff>=0.1.0
mypy>=1.0.0
httpx>=0.24.0

# Web Admin dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
jinja2>=3.1.0
python-multipart>=0.0.6
itsdangerous>=2.1.0
